LaneKeeping3/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.1.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
pillow>=9.5.0

# Data processing
pandas>=2.0.0
h5py>=3.8.0
pycocotools>=2.0.6

# Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Tracking & filtering
filterpy>=1.4.5

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Logging & monitoring
tensorboard>=2.12.0
wandb>=0.15.0
mlflow>=2.3.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0

# Deployment
onnx>=1.14.0
onnxruntime>=1.15.0
# tensorrt>=8.6.0  # Install separately based on CUDA version

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0
